pip-requirements.txt
Django==1.8.5
django-localflavor==1.1
psycopg2==2.6.1
